pyproject.toml
[project]
name = "ai-cv-generator"
version = "0.1.0"
description = "Generate and refine CSs using LLMs."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
    { name = "Peter Kaszt", email = "kasztp@gmail.com" },
]
dependencies = [
    "google-generativeai>=0.5.0", # Check for the latest version
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.4.0", # Use uv's built-in ruff or install explicitly
]

[project.scripts]
generate-cv = "cv_generator.generator:main"

# Tool Configurations
[tool.uv]
# Optional: If you want uv to manage Ruff/Pytest versions explicitly
# requires = ["ruff", "pytest"]

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []
# fixable = ["ALL"] # Uncomment to enable auto-fixing for all rules

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath = ["src"] # Make src importable in tests
